+++ changelog.md
@@ -11,3 +11,8 @@ Initial SS4 release
 ## 2.0.1
 
 Update config inline with latest SS
+
+## 2.0.2
+
+ * Add additional validation to ensure a member has EITHER a primary ident or alt ident
+ * Remove requirement for member to have an email address